engine: extract esm/imageset.js
