+++ engine/esm/data_globals.js
@@ -12,3 +12,11 @@ export var freestandingMode = false;
 export function set_freestandingMode(value) {
     freestandingMode = !!value;
 }
+
+// This is equivalent to `new HipsProperties(imageset)`. We abstract
+// the function to avoid circular dependencies in the type hierarchy.
+export var makeNewHipsProperties = null;
+
+export function set_makeNewHipsProperties(value) {
+    makeNewHipsProperties = value;
+}
